fix(rust): fix countdown loop underflow and remaining runtime issues
Five fixes that take Rust regression tests from 20/163 to 161/163:

1. Countdown loop underflow: for(v=N; v>=bound; v--) with usize wraps
   to usize::MAX when v reaches bound. Emit loop-with-break-at-bound
   pattern instead of while loop.

2. Signed integer detection: int variables used with negative values
   (assigned 0-N, compared < 0) now emit as i32 instead of usize.
   Fixes CDLHIKKAKE and similar candlestick patterns.

3. Inlined helper precedence: FuncCall inlined to BinOp now wraps in
   parens to preserve operator precedence (e.g., highlowrange(H,L)/2
   must be (H-L)/2, not H-L/2). Fixes CDLRICKSHAWMAN and ~20 others.

4. ForC range-iteration post-loop fixup: after for-in loop, set
   iterator to end+1 to match C post-loop semantics. Fixes CORREL
   and ~15 others that use the loop variable after the for loop.

5. Nested VarDecl init expressions: VarDecl inside loops now renders
   with its init expression instead of zero. Fixes IMI.

+++ ta_codegen_output/rust/src/ta_func/nvi.rs
@@ -53,8 +53,8 @@ impl Core {
     ///
     /// # Arguments
     ///
-    pub fn nvi_lookback(&self) -> i32 {
-        return 0;
+    pub fn nvi_lookback(&self) -> usize {
+        return (0) as usize;
     }
     /// Negative Volume Index
     ///
